tests(chore): clean up tags
see `pytest.ini` to get overview of supported tags
