+++ e2e-tests/pytest.ini
@@ -1,22 +1,23 @@
 [pytest]
 markers =
-    CD: run tests meant for Continuous Deployment workflow
+    # smoke tests
+    smoke: block producing, transactions, etc.
+    rpc: node RPC methods
+
+    # main features
+    ariadne: committee selection algorithm
+    reserve: reserve management system
+    governed_map: arbitrary data stored on the main chain
+    delegator_rewards: ADA delegator rewards
+    
+    # helper tags
     skip_on_new_chain: skip test on new chain (less than 2 MC epochs have passed)
     candidate_status: active or inactive, used in test_registrations.py
-    ariadne: run tests for Ariadne feature
-    probability: run non deterministic tests
-    delegator_rewards: run tests for ADA Delegator Rewards feature
-    registration: run tests for Registration feature (permissioned and trustless)
-    committee_distribution: run tests for Committee Distribution feature
-    committee_rotation: run tests for Committee Rotation feature
-    committee_members: run tests for Committee Members feature
-    mc_state_reference_block: run tests for Mainchain State Reference feature
-    block_production_log: run test for block production log pallet
-    reserve: run tests for Reserve Managemenet System feature
-    governed_map: run tests for Governed Map feature
-    rpc: run tests for RPC endpoints
-    substrate: run tests for Substrate framework core features
-    test_key: key for xray
+
+    # deprecated tags - soon to be removed
+    probability: (deprecated) used to skip committee participation on short epochs chains
+    test_key: (deprecated) xray id
+
 junit_duration_report = call
 log_format = %(asctime)s %(levelname)-8s [logger=%(name)s, file=%(filename)s:%(lineno)d] %(message)s
 log_date_format = %Y-%m-%d %H:%M:%S %Z
